Test & Refactor Test, delete some test file and re organized new ones Provide test to ensure BaseRenderer Behaviour, tests file contains multiple desc, each one cover a single method purpose
